Update bvseosdk.php
Updated the construct method to correctly use the current_page_url value that's sent to the script.  Updated _getPageNumber method such that when the url is like this... 

.../example.php?bvrrp=12325/reviews/product/1/product1.htm

the page number derived from the above url is used versus the url that's defined in the current_page_url variable.  example:
    $bv = new BV(array(
        'deployment_zone_id' => 'Main_Site-en_US',
        'product_id' => 'PRTG2000',
        'cloud_key' => 'totalgymdirect-f19e9cfeb9917d0c20afdfbe06483f1b',
        'staging' => FALSE,
        'current_page_url' => 'bvrrp=12325/reviews/product/2/product1.htm',
    ));